+++ CHANGELOG.md
@@ -8,6 +8,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.15.0] - 2024-12-23
+
+### Added
+
+- `KPIconButton` and `KPContainedIconButton` APIs introduced
+- Added Pause, Play, Save, Share icons
+
 ## [0.14.0] - 2024-09-10
 
 ### Added